feat: implement authentication module with JWT support and route protection

+++ cookops-backend/src/assinatura/assinatura.controller.ts
@@ -2,46 +2,73 @@ import {
   Body,
   Controller,
   Delete,
+  ForbiddenException,
   Get,
   Param,
   Post,
   Put,
+  Request,
+  UseGuards,
 } from '@nestjs/common';
+import { JwtAuthGuard } from 'src/auth/jwt/jwt-auth.guard';
 import { AssinaturaService } from './assinatura.service';
 import { CreateAssinaturaDto } from './dto/create-assinatura.dto';
 import { UpdateAssinaturaDto } from './dto/update-assinatura.dto';
 
 @Controller('assinaturas')
+@UseGuards(JwtAuthGuard)
 export class AssinaturaController {
   constructor(private readonly assinaturaService: AssinaturaService) {}
 
   @Post()
-  create(@Body() data: CreateAssinaturaDto) {
+  create(
+    @Request() req: { user: { role: string } },
+    @Body() data: CreateAssinaturaDto,
+  ) {
+    if (req.user.role !== 'ADMIN')
+      throw new ForbiddenException('Acesso negado');
     return this.assinaturaService.create(data);
   }
 
   @Get()
-  findAll() {
+  findAll(@Request() req: { user: { role: string } }) {
+    if (req.user.role !== 'ADMIN')
+      throw new ForbiddenException('Acesso negado');
     return this.assinaturaService.findAll();
   }
 
   @Get(':id')
-  findOne(@Param('id') id: string) {
+  findOne(@Request() req: { user: { role: string } }, @Param('id') id: string) {
+    if (req.user.role !== 'ADMIN')
+      throw new ForbiddenException('Acesso negado');
     return this.assinaturaService.findOne(id);
   }
 
   @Put(':id')
-  update(@Param('id') id: string, @Body() data: UpdateAssinaturaDto) {
+  update(
+    @Request() req: { user: { role: string } },
+    @Param('id') id: string,
+    @Body() data: UpdateAssinaturaDto,
+  ) {
+    if (req.user.role !== 'ADMIN')
+      throw new ForbiddenException('Acesso negado');
     return this.assinaturaService.update(id, data);
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
+  remove(@Request() req: { user: { role: string } }, @Param('id') id: string) {
+    if (req.user.role !== 'ADMIN')
+      throw new ForbiddenException('Acesso negado');
     return this.assinaturaService.remove(id);
   }
 
   @Get('empresa/:empresaId')
-  findByEmpresaId(@Param('empresaId') empresaId: string) {
+  findByEmpresaId(
+    @Request() req: { user: { role: string } },
+    @Param('empresaId') empresaId: string,
+  ) {
+    if (req.user.role !== 'ADMIN')
+      throw new ForbiddenException('Acesso negado');
     return this.assinaturaService.findByEmpresaId(empresaId);
   }
 }
