+++ cookops-backend/src/empresa/empresa.controller.ts
@@ -2,41 +2,63 @@ import {
   Body,
   Controller,
   Delete,
+  ForbiddenException,
   Get,
   Param,
   Post,
   Put,
+  Request,
+  UseGuards,
 } from '@nestjs/common';
+import { JwtAuthGuard } from 'src/auth/jwt/jwt-auth.guard';
 import { CreateEmpresaDto } from './dto/create-empresa.dto';
 import { UpdateEmpresaDto } from './dto/update-empresa.dto';
 import { EmpresaService } from './empresa.service';
 
 @Controller('empresas')
+@UseGuards(JwtAuthGuard)
 export class EmpresaController {
   constructor(private readonly empresaService: EmpresaService) {}
 
   @Post()
-  create(@Body() data: CreateEmpresaDto) {
+  create(
+    @Request() req: { user: { role: string } },
+    @Body() data: CreateEmpresaDto,
+  ) {
+    if (req.user.role !== 'ADMIN')
+      throw new ForbiddenException('Acesso negado');
     return this.empresaService.create(data);
   }
 
   @Get()
-  findAll() {
+  findAll(@Request() req: { user: { role: string } }) {
+    if (req.user.role !== 'ADMIN')
+      throw new ForbiddenException('Acesso negado');
     return this.empresaService.findAll();
   }
 
   @Get(':id')
-  findOne(@Param('id') id: string) {
+  findOne(@Request() req: { user: { role: string } }, @Param('id') id: string) {
+    if (req.user.role !== 'ADMIN')
+      throw new ForbiddenException('Acesso negado');
     return this.empresaService.findOne(id);
   }
 
   @Put(':id')
-  update(@Param('id') id: string, @Body() data: UpdateEmpresaDto) {
+  update(
+    @Request() req: { user: { role: string } },
+    @Param('id') id: string,
+    @Body() data: UpdateEmpresaDto,
+  ) {
+    if (req.user.role !== 'ADMIN')
+      throw new ForbiddenException('Acesso negado');
     return this.empresaService.update(id, data);
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
+  remove(@Request() req: { user: { role: string } }, @Param('id') id: string) {
+    if (req.user.role !== 'ADMIN')
+      throw new ForbiddenException('Acesso negado');
     return this.empresaService.remove(id);
   }
 }
